feat(tasks): upload stats to discords.com
